refactor(shared): promote JsonTree to a first-class peer component
Move the JSON viewer out of components/trace/ into its own peer at
components/json-tree/ so consumers no longer reach into the tracer for a
general-purpose viewer.

- git-mv trace/JsonTree.tsx -> json-tree/JsonTree.tsx (history preserved)
- add json-tree/index.ts barrel; export IJsonTreeProps
- repoint trace's internal consumers (TraceDetail + 4 renderers) to ../json-tree
- drop the JsonTree re-export from the trace barrel (no external consumer used it)

No behavior change. Verified: builder shared:test 77/77 (incl. the JsonTree
5MB-cap test through the new path).
