+++ apps/shared/src/components/json-tree/index.ts
@@ -1,0 +1,2 @@
+export { JsonTree } from './JsonTree';
+export type { IJsonTreeProps } from './JsonTree';
