+++ apps/shared/src/components/trace/index.tsx
@@ -1,4 +1,3 @@
 export { default as Trace } from './Trace';
 export { TraceDetail } from './TraceDetail';
 export type { ITraceDetailProps, ITraceFetchResult } from './TraceDetail';
-export { JsonTree } from './JsonTree';
